Cleanup MergeSegments and GetSegments API
