+++ rs/proto/proto/muopdb.proto
@@ -45,36 +45,7 @@ service IndexServer {
   rpc InsertPacked(InsertPackedRequest) returns (InsertPackedResponse) {}
 
   rpc Flush(FlushRequest) returns (FlushResponse) {}
-
-  rpc GetSegments(GetSegmentsRequest) returns (GetSegmentsResponse) {}
-
-  rpc CompactSegments(CompactSegmentsRequest) returns (CompactSegmentsResponse) {}
 }
-
-message GetSegmentsRequest {
-  string collection_name = 1;
-}
-
-message SegmentInfo {
-  string segment_name = 1;
-  uint64 size_in_bytes = 2;
-}
-
-message GetSegmentsResponse {
-  // Deprecated: Use segment_infos instead
-  repeated string segment_names = 1;
-
-  repeated SegmentInfo segment_infos = 2;
-}
-
-message CompactSegmentsRequest {
-  string collection_name = 1;
-  repeated string segment_names = 2;
-}
-
-message CompactSegmentsResponse {
-}
-
 
 message CreateCollectionRequest {
   string collection_name = 1;
